requirements.txt
requests
azure-data-tables
curl_cffi
dotenv
pandas
openpyxl
# dateutil